requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# ── Database ──────────────────────────────────────────────────────────────────
duckdb==0.10.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3==1.34.113
s3fs==2024.3.1

# ── dbt ───────────────────────────────────────────────────────────────────────
dbt-core==1.8.3
dbt-redshift==1.8.0
dbt-duckdb==1.8.1          # for local dev / CI without Redshift

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.1
lifetimes==0.11.3           # BG/NBD + Gamma-Gamma LTV models
mlflow==2.13.2
optuna==3.6.1               # hyperparameter tuning

# ── Jupyter ───────────────────────────────────────────────────────────────────
jupyterlab==4.2.1
ipywidgets==8.1.3
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.35.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.1
pytest-cov==5.0.0

# ── Code quality ──────────────────────────────────────────────────────────────
ruff==0.4.7
